+++ source/main/utils/makeEnvironmentGlobal.js
@@ -1,6 +1,0 @@
-/**
- * Make the main process environment accessible to renderers
-*/
-export const makeEnvironmentGlobal = (env) => {
-  Object.assign(global, { env });
-};