+++ source/renderer/app/api/api.js
@@ -46,13 +46,9 @@ import { updateWallet } from './wallets/requests/updateWallet';
 // utility functions
 import { awaitUpdateChannel, cardanoFaultInjectionChannel } from '../ipc/cardano.ipc';
 import patchAdaApi from './utils/patchAdaApi';
-import { isValidMnemonic } from '../../../common/decrypt';
+import { isValidMnemonic } from '../../../common/crypto/decrypt';
 import { utcStringToDate, encryptPassphrase } from './utils';
-import {
-  Logger,
-  stringifyData,
-  stringifyError
-} from '../../../common/logging';
+import { Logger } from '../utils/logging';
 import {
   isValidRedemptionKey,
   isValidPaperVendRedemptionKey
@@ -153,7 +149,8 @@ import {
   NotEnoughMoneyToSendError,
   RedeemAdaError
 } from './transactions/errors';
-import type { FaultInjectionIpcRequest } from '../../../common/types/cardanoNode.types';
+import type { FaultInjectionIpcRequest } from '../../../common/types/cardano-node.types';
+import { stringifyData, stringifyError } from '../../../common/utils/logging';
 
 export default class AdaApi {
 
@@ -622,10 +619,10 @@ export default class AdaApi {
     }
   }
 
-  async sendBugReport(requestFormData: SendBugReportRequest): Promise<any> {
-    Logger.debug('AdaApi::sendBugReport called: ' + stringifyData(requestFormData));
+  async sendBugReport({ requestFormData, environmentData }: SendBugReportRequest): Promise<any> {
+    Logger.debug('AdaApi::sendBugReport called: ' + stringifyData({ requestFormData, environmentData }));
     try {
-      await sendBugReport({ requestFormData });
+      await sendBugReport({ requestFormData, environmentData });
       Logger.debug('AdaApi::sendBugReport success');
       return true;
     } catch (error) {