+++ source/renderer/app/components/wallet/paper-wallet-certificate/CompletionDialog.js
@@ -11,6 +11,7 @@ import { getNetworkExplorerUrl } from '../../../utils/network';
 import styles from './CompletionDialog.scss';
 import iconCopy from '../../../assets/images/clipboard-ic.inline.svg';
 import InlineNotification from '../../widgets/InlineNotification';
+import { DEVELOPMENT } from '../../../../../common/types/environment.types';
 
 const messages = defineMessages({
   headline: {
@@ -61,6 +62,7 @@ type Props = {
   onClose: Function,
   onOpenExternalLink: Function,
   copyAddressNotificationDuration: number,
+  network: string
 };
 
 type State = {
@@ -72,6 +74,10 @@ export default class CompletionDialog extends Component<Props, State> {
 
   static contextTypes = {
     intl: intlShape.isRequired,
+  };
+
+  static defaultProps = {
+    network: DEVELOPMENT
   };
 
   state = {
@@ -92,7 +98,7 @@ export default class CompletionDialog extends Component<Props, State> {
 
   render() {
     const { intl } = this.context;
-    const { onClose, walletCertificateAddress, onOpenExternalLink } = this.props;
+    const { onClose, walletCertificateAddress, onOpenExternalLink, network } = this.props;
     const { showCopyNotification } = this.state;
     const dialogClasses = classnames([
       styles.component,
@@ -107,7 +113,7 @@ export default class CompletionDialog extends Component<Props, State> {
         onClick: onClose,
       }
     ];
-    const cardanoExplorerLink = `${getNetworkExplorerUrl()}/address/${walletCertificateAddress}`;
+    const cardanoExplorerLink = `${getNetworkExplorerUrl(network)}/address/${walletCertificateAddress}`;
 
     // Get QRCode color value from active theme's CSS variable
     const qrCodeBackgroundColor = document.documentElement ?